# z: use zeebo/xxh3 instead of cespare/xxhash/v2