--- a/go.mod
+++ b/go.mod
@@ -3,12 +3,12 @@ module github.com/dgraph-io/ristretto
 go 1.12
 
 require (
-	github.com/cespare/xxhash/v2 v2.1.1
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/dgryski/go-farm v0.0.0-20190423205320-6a90982ecee2
 	github.com/dustin/go-humanize v1.0.0
 	github.com/golang/glog v0.0.0-20160126235308-23def4e6c14b
 	github.com/pkg/errors v0.9.1
 	github.com/stretchr/testify v1.4.0
+	github.com/zeebo/xxh3 v1.0.1
 	golang.org/x/sys v0.0.0-20200930185726-fdedc70b468f
 )