--- a/z/z.go
+++ b/z/z.go
@@ -20,7 +20,7 @@ import (
 	"context"
 	"sync"
 
-	"github.com/cespare/xxhash/v2"
+	"github.com/zeebo/xxh3"
 )
 
 // TODO: Figure out a way to re-use memhash for the second uint64 hash, we
@@ -39,9 +39,9 @@ func KeyToHash(key interface{}) (uint64, uint64) {
 	case uint64:
 		return k, 0
 	case string:
-		return MemHashString(k), xxhash.Sum64String(k)
+		return MemHashString(k), xxh3.HashString(k)
 	case []byte:
-		return MemHash(k), xxhash.Sum64(k)
+		return MemHash(k), xxh3.Hash(k)
 	case byte:
 		return uint64(k), 0
 	case int: