Use zeebo/xxh3 instead of cespare/xxhash/v2.

```
darwin/amd64 (might be disabled AVX512 even if Xeon, cuz darwin kernel bug)
name                                       old time/op    new time/op    delta
KeyToHashBytes/foo-20                        9.08ns ± 1%    7.96ns ± 1%  -12.39%  (p=0.008 n=5+5)
KeyToHashBytes/barbaz-20                     9.35ns ± 0%    8.05ns ± 2%  -13.91%  (p=0.008 n=5+5)
KeyToHashBytes/quxquuxquuz-20                10.1ns ± 0%     8.1ns ± 4%  -19.69%  (p=0.008 n=5+5)
KeyToHashBytes/corgegraultgarplywaldo-20     11.0ns ± 4%     9.4ns ± 3%  -14.64%  (p=0.008 n=5+5)
KeyToHashString/foo-20                       11.1ns ± 1%     7.5ns ± 3%  -32.18%  (p=0.008 n=5+5)
KeyToHashString/barbaz-20                    11.2ns ± 2%     7.7ns ± 1%  -31.48%  (p=0.008 n=5+5)
KeyToHashString/quxquuxquuz-20               12.3ns ± 0%     7.5ns ± 1%  -38.90%  (p=0.008 n=5+5)
KeyToHashString/corgegraultgarplywaldo-20    13.7ns ± 0%     8.6ns ± 0%  -36.79%  (p=0.008 n=5+5)

name                                       old speed      new speed      delta
KeyToHashBytes/foo-20                       330MB/s ± 1%   377MB/s ± 1%  +14.14%  (p=0.008 n=5+5)
KeyToHashBytes/barbaz-20                    642MB/s ± 0%   746MB/s ± 2%  +16.17%  (p=0.008 n=5+5)
KeyToHashBytes/quxquuxquuz-20              1.09GB/s ± 0%  1.36GB/s ± 4%  +24.60%  (p=0.008 n=5+5)
KeyToHashBytes/corgegraultgarplywaldo-20   1.99GB/s ± 4%  2.33GB/s ± 3%  +17.12%  (p=0.008 n=5+5)
KeyToHashString/foo-20                      271MB/s ± 1%   399MB/s ± 3%  +47.53%  (p=0.008 n=5+5)
KeyToHashString/barbaz-20                   537MB/s ± 2%   784MB/s ± 1%  +45.93%  (p=0.008 n=5+5)
KeyToHashString/quxquuxquuz-20              894MB/s ± 0%  1463MB/s ± 1%  +63.67%  (p=0.008 n=5+5)
KeyToHashString/corgegraultgarplywaldo-20  1.61GB/s ± 0%  2.55GB/s ± 0%  +58.21%  (p=0.008 n=5+5)

linux/amd64 (enabled AVX512)
name                                       old time/op    new time/op    delta
KeyToHashBytes/foo-30                        10.4ns ± 0%     8.6ns ± 0%  -17.72%  (p=0.008 n=5+5)
KeyToHashBytes/barbaz-30                     10.4ns ± 0%     8.9ns ± 0%  -14.69%  (p=0.016 n=4+5)
KeyToHashBytes/quxquuxquuz-30                11.6ns ± 0%     8.9ns ± 1%  -23.70%  (p=0.008 n=5+5)
KeyToHashBytes/corgegraultgarplywaldo-30     12.6ns ± 0%    10.5ns ± 0%  -16.68%  (p=0.008 n=5+5)
KeyToHashString/foo-30                       12.9ns ± 0%     8.3ns ± 0%  -35.94%  (p=0.008 n=5+5)
KeyToHashString/barbaz-30                    12.8ns ± 1%     8.6ns ± 1%  -32.54%  (p=0.008 n=5+5)
KeyToHashString/quxquuxquuz-30               14.2ns ± 0%     8.9ns ± 0%  -37.52%  (p=0.016 n=4+5)
KeyToHashString/corgegraultgarplywaldo-30    15.7ns ± 0%    10.0ns ± 1%  -36.75%  (p=0.008 n=5+5)

name                                       old speed      new speed      delta
KeyToHashBytes/foo-30                       288MB/s ± 0%   350MB/s ± 0%  +21.53%  (p=0.008 n=5+5)
KeyToHashBytes/barbaz-30                    577MB/s ± 0%   676MB/s ± 0%  +17.15%  (p=0.008 n=5+5)
KeyToHashBytes/quxquuxquuz-30               948MB/s ± 0%  1242MB/s ± 1%  +31.10%  (p=0.008 n=5+5)
KeyToHashBytes/corgegraultgarplywaldo-30   1.75GB/s ± 0%  2.10GB/s ± 1%  +20.02%  (p=0.008 n=5+5)
KeyToHashString/foo-30                      232MB/s ± 0%   362MB/s ± 0%  +56.08%  (p=0.008 n=5+5)
KeyToHashString/barbaz-30                   469MB/s ± 1%   696MB/s ± 1%  +48.22%  (p=0.008 n=5+5)
KeyToHashString/quxquuxquuz-30              776MB/s ± 0%  1242MB/s ± 0%  +60.09%  (p=0.008 n=5+5)
KeyToHashString/corgegraultgarplywaldo-30  1.40GB/s ± 0%  2.21GB/s ± 1%  +58.09%  (p=0.008 n=5+5)
```

benchmark result:
- https://perf.golang.org/search?q=upload:20211112.4
- https://perf.golang.org/search?q=upload:20211112.5

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/dgraph-io/ristretto/291)
<!-- Reviewable:end -->